[modify] add http trigger
